+++ packages/db/src/drivers/mysql.js
@@ -12,6 +12,11 @@ module.exports.createPool = (dbconfig) => {
   return mysql.createPool(_.pick(dbconfig, OPTIONS));
 };
 
+// close pool
+module.exports.closePool = async (pool) => {
+  await pool.end();
+};
+
 // get connection
 module.exports.getConnection = async (pool) => {
   return await pool.getConnection();